Add homebrew and updated the "-up" command and builder.sh
On Dockerfile was added homebrew, tmux and neovim to the container.
On "-up" command, added the support to snap, flatpak and brew update
On the builder.sh (install function) fixed up the installer if already had an older version of the script installed.

Container with homebrew working

Add tmux and nvim on debug container, updated the update function and
the builder install